return generic value (instead of Object) for Action.getValue(index)

+++ lib/src/test/java/com/yheriatovych/reductor/CursorsTest.java
@@ -71,8 +71,8 @@ public class CursorsTest {
             }
             if(action.type.equals("UPDATE")) {
                 ComplexState nextState = new ComplexState();
-                nextState.foo = (int) action.getValue(0);
-                nextState.bar = (int) action.getValue(1);
+                nextState.foo = action.getValue(0);
+                nextState.bar = action.getValue(1);
                 return nextState;
             }
             return state;
